Adjust for consistent widget layout on Jelly Bean.
The widget was displaying running chronometer text right justified and
paused text left justified. The fix is to make the horizontal LinearLayout
the place to do gravity="right".

Also it was resizing the start/stop button. The workaround is to explicitly
set the ViewFlipper to measureAllChildren="false" even though that's the
default.

Also turn on ellipsize="end". Does it work?

Add contentDescription to the start/stop ImageButton for accessibility.