An experiment in putting a spatula image behind the widget.
Punted. `android:tileMode="clamp"` will replicate the edges but it
replicates the edges of a trimmed image for some reason, so the tip of the
spatula repeats across the image to the right. Also tileMode defeats the
android:gravity setting so the image sticks in the upper left corner.
`android:tileMode="repeat"` is too busy.

Stash the spatula image files and its background drawable XML file in
source-assets/.