Fix the checkbox layout (overlaps its label)
Setting its padding (in particular its paddingLeft) overrides the
default padding which leaves barely enough room for the checkbox in API
levels < 17, and begins at the edge of the checkbox in API levels >= 17.
One could write a workaround that conditionally increases the checkbox's
left padding, but the default is good enough as long as we don't
override it with a fixed value.

See
http://stackoverflow.com/questions/4037795/android-spacing-between-checkbox-and-text/4038195#4038195

versionName "1.0 beta 2", versionCode 8