Improve the TAB behavior in MainActivity, but it's very buggy.
Even after a lot of experiments with these fragile attributes, the OS's
behavior (at least in the emulators) is buggy.

* Focus on the timer text doesn't show any signifier.
* Ditto for the check box in some OS versions and sometimes depending on
whether it's checked on. (This might be due to the Material Design theme.)
* TAB focus goes to the Reset button even when the button is hidden.
* TAB focus gets stuck at the number picker (the last widget in the
activity) in the API level 21 emulator. This makes focus on the number
picker useless since you can TAB there, type a number, and have no way to
make it accept the new number.
* The widgets need explicit android:focusable="true".
* The Reset button needs an explicit android:nextFocusForward attribute or
else it takes two TABs to get to the next widget.
* The arrow keys do nothing.
* In the API 16-17 emulators, TAB does nothing.
* In the API 15 & 12 emulators, TAB sometimes goes to the number picker's
UP or DOWN arrow, depending on whether the Reset button is visible and
whether the number picker is set to 1.