use AlarmManager.setAlarmClock() to wake from idle/doze
This should cope with Android Marshmallow’s battery optimizing
features, but Developer Preview 3 is inadequate for testing.

Calling setAlarmClock() displays another timer icon in the
notification bar and lock screen with tap through to MainActivity
to edit the timer. It’s available on Lollipop+ so use it there.

setAlarmClock() uses RTC wall time instead of elapsed
interval time, so convert time bases and listen for clock
adjustments.

Hopefully it’s a UI improvement, but alas it displays the alarm time
as an absolute minute rather than an interval, and it could seem
broken since the alarm doesn’t happen at the turn of the minute.