Pre-Lollipop, unify Paused @ 0:00 with Stopped in the UI
For Android L+, the app has Stop *and* Reset (go to Paused @ 0:00)
buttons.

For older Androids, the app only needs one of them. This release
switches it from Reset to Stop.

The difference now matters because the Activity pulses the time display
when Paused, thus upgrading from an older app release needs to change
the stored Reset state to Stopped.