small count-up text for 2x1 size widgets
`setTextViewTextSize()` causes bugs and workarounds were only partially successful.

What does work is to switch to separately sized children in viewFlipper. It's complicated.

Implications are hard to predict. Loading layout resources each time does not seem to start fresh. Perhaps a cache doesn't get cleared properly. So it's important to stop the chronometers when not needed, set the text of the `smallPausedChronometerText` field even when hidden, etc.