more intuitive time parser
It needn't be clever with spaces since the EditText with inputType="time" rejects spaces, even from a physical keyboard.

So just remove all spaces then treat each ":" (rather than a run of colons and spaces) as a field separator. E.g. "::" is "0:0:0" and ":::" is invalid.

Explain the interval time format in the Play Store description.