Shrink German full_description.txt
Fix #9

Shrink the German app description to fit Fastlane, which is 4000 characters after the tools do HTML conversion (and prepends the app name or the app short description?).

That's a bit hard to predict. Izzy asked to reduce it by ≈81 chars. This PR reduces it by 169 chars using Izzy's recommendations and trimming fine details in German and English. Hopefully the German translation is good.