Roboto font for fixed-width digits on Moto G
Moto G's built-in fonts (are they all the same?) have variable-width digits, which makes the timer display jittery, esp. the rapidly changing tenths of a second.

Workaround: Include the Roboto (regular weight) font and use that.

Unfortunately a custom font doesn't work in the app widget and the notification shade.