show a "phone is locked" icon in the action bar
It's surprising that the app can open from the lock screen notification or widget without unlocking, also that turning on the screen while the phone is locked will show the app if it was visible when the screen turned off.

The Android docs say the app is open "on top of the lock screen," but this is not revealed to users.

Provide feedback via a "phone is (still) locked" icon in the action bar so the user doesn't worry about security.